+++ datadog_checks_dev/datadog_checks/dev/tooling/cli.py
@@ -6,7 +6,7 @@ import os
 import click
 
 from .commands import ALL_COMMANDS
-from .commands.utils import CONTEXT_SETTINGS, echo_success, echo_waiting, echo_warning
+from .commands.console import CONTEXT_SETTINGS, echo_success, echo_waiting, echo_warning
 from .config import CONFIG_FILE, config_file_exists, load_config, restore_config
 from .constants import set_root
 from ..compat import PermissionError
