+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/create.py
@@ -6,7 +6,7 @@ from collections import defaultdict
 
 import click
 
-from .utils import CONTEXT_SETTINGS, abort, echo_info, echo_success
+from .console import CONTEXT_SETTINGS, abort, echo_info, echo_success
 from ..constants import get_root
 from ..create import construct_template_fields, create_template_files, get_valid_templates
 from ..utils import normalize_package_name
