+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/validate/dep.py
@@ -6,7 +6,7 @@ import os
 import click
 from six import iteritems
 
-from ..utils import CONTEXT_SETTINGS, abort, echo_failure, echo_info
+from ..console import CONTEXT_SETTINGS, abort, echo_failure, echo_info
 from ...constants import get_root, get_agent_requirements
 from ...dep import make_catalog, read_packages
