+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/validate/service_checks.py
@@ -8,7 +8,7 @@ from collections import OrderedDict
 import click
 from six import string_types
 
-from ..utils import CONTEXT_SETTINGS, abort, echo_failure, echo_info, echo_success
+from ..console import CONTEXT_SETTINGS, abort, echo_failure, echo_info, echo_success
 from ...constants import get_root
 from ...utils import parse_version_parts
 from ....compat import JSONDecodeError
